Improve dynamic analyser

+++ sandbox/src/app.ts
@@ -9,23 +9,39 @@ import yaml from 'yaml';
 import { sha256 } from 'js-sha256';
 import * as log from "log4js";
 import { validate_logging_level } from "./utils";
-const args = process.argv;
+import { default_rabbitmq_conf } from "./const"
+import commandLineArgs from 'command-line-args'
 
-var sampleFile = args[2]
-var baitWebsite = args[3]
-var configFolder = args[4]
-var analysisId = args[5]
-
+const MAX_SET_TIMEOUT_DELAY_TO_WAIT = 5000;
 const log_level = process.env.LOG_LEVEL;
 let logger = log.getLogger();
 logger.level = validate_logging_level(log_level ? log_level : 'info')
+// const args = process.argv;
+const cli_arg_options = [
+    { name: 'dev', alias: 'd', type: Boolean, defaultValue: false },
+    { name: 'just-check-page', alias: 'p', type: Boolean, defaultValue: false },
+    { name: 'sample-file', alias: 'f', type: String, defaultValue: "" },
+    { name: 'bait-website', alias: 'w', type: String, defaultValue: "https://google.com" },
+    { name: 'conf-folder', alias: 'c', type: String, defaultValue: "/sandbox/config" },
+    { name: 'analysis-id', alias: 'a', type: String, defaultValue: ""}
+]
 
-if (!sampleFile) {
+const cli_args = commandLineArgs(cli_arg_options)
+
+logger.debug("cli_args: ", cli_args)
+var sampleFile = cli_args["sample-file"]
+var justCheckPage = cli_args["just-check-page"]
+
+var baitWebsite = cli_args["bait-website"]
+var configFolder = cli_args["conf-folder"]
+var analysisId = cli_args["analysis-id"]
+
+if (!sampleFile && !justCheckPage) {
     logger.error("Error: No JavaScript file provided!");
     process.exit(1);
 }
 
-if (!fs.existsSync(sampleFile)) {
+if (!fs.existsSync(sampleFile) && !justCheckPage) {
     logger.error(`Error: File "${sampleFile}" not found!`);
     process.exit(1);
 }
@@ -44,23 +60,65 @@ if (!fs.existsSync(sampleFile)) {
     });
     logger.info("[analysis-info] Connecting to RabbitMQ ...");
 
-    const rabbitmq_config_file = fs.readFileSync(configFolder + "/rabbitmq.yaml", 'utf8');
-    const rabbitmq_config: types.RabbitMQConfig = yaml.parse(rabbitmq_config_file);
-
-    let rbmqc = await RBMQ.create(rabbitmq_config)
+    let rbmqc = await RBMQ.empty()
+    let rabbitmq_config: types.RabbitMQConfig = default_rabbitmq_conf;
+    try {
+        const rabbitmq_config_file = fs.readFileSync(configFolder + "/rabbitmq.yaml", 'utf8');
+        rabbitmq_config = yaml.parse(rabbitmq_config_file);
+        rbmqc = await RBMQ.create(rabbitmq_config)
+    } catch(err) {
+        if (cli_args.dev) {
+            logger.warn("[analysis-warn] In development mode, skipping queue connection ...");
+        } else {
+            logger.error(`[analysis-error] Could not connect to queue: ${err}`);
+            process.exit(1)
+        }
+    }
+    let iocs: types.IoC[] = [];
 
     logger.info("[analysis-info] Launched ...");
     const page = await browser.newPage();
+    await page.setRequestInterception(true);
+      page.on('request', request => {
+        if(request.url().endsWith('.zip')) {
+            logger.debug('[analysis-debug] adding ioc: downloads zip file');
+            iocs.push(
+                {
+                    type: types.IoCType.HttpRequest,
+                    timestamp: Date.now(),
+                    executed_on: baitWebsite,
+                    value: {
+                        url: request.url(),
+                        method:"GET",
+                        data: ""
+                    } as types.IoCHttpRequest
+                }
+            );
+        }
+        request.continue();
+    });
+    // page.on('response', async response => {
+    //     // Continue all requests
+    //     if (response.request() && response.status() < 300 && response.status() >= 400) {
+    //         var buffer = await response.buffer(); /*You can get the buffer*/
+    //         var content = await response.text(); /*You can get the content as text*/
+    //     }
+    // });
+    browser.on('targetcreated', async target => {
+        logger.warn("new page loaded: ", target)
+    })
 
     // visiting bait_websites
-    await page.goto(baitWebsite);
+    await page.goto(baitWebsite, {
+        waitUntil: ['domcontentloaded', 'networkidle0']
+    });
     logger.info("[analysis-info] Hooking JavaScript APIs...");
 
     page.on('console', message => {
         logger.debug(`[dom-console]: ${message.text()}`);
     })
 
-    let iocs: types.IoC[] = [];
+    let max_delay = 0;
 
     // a random string suffix is created every time the analyser runs to make
     // sure Javascript cannot find and call our hook and mess with the analyser
@@ -71,11 +129,21 @@ if (!fs.existsSync(sampleFile)) {
     // here we actually push iocs to the their final destination, the sandbox_ioc_queue
     // we can perform any kind of normalization to the event
     await page.exposeFunction(reportIocFunctionName, (ioc: types.IoC) => {
-        logger.debug('[analysis-debug] adding ioc: ', ioc);
-        ioc.executed_on = baitWebsite
-        iocs.push(
-            ioc
-        );
+        if (ioc.type == types.IoCType.SetTimeout) {
+            let setTimeoutDelay = (ioc.value as types.IoCSetTimeout).delay
+            logger.debug(
+                '[analysis-debug] setTimeout was set with delay: ',
+                setTimeoutDelay)
+                if (max_delay < setTimeoutDelay && setTimeoutDelay <= MAX_SET_TIMEOUT_DELAY_TO_WAIT) {
+                    max_delay = setTimeoutDelay
+                }
+        } else {
+            logger.debug('[analysis-debug] adding ioc: ', ioc);
+            ioc.executed_on = baitWebsite
+            iocs.push(
+                ioc
+            );
+        }
     });
     const place_hooks_source_code = place_hooks.toString();
 
@@ -92,10 +160,15 @@ if (!fs.existsSync(sampleFile)) {
         await page.evaluate(place_hooks_wrapper);
         logger.info(`[analysis-info] Executing sample`);
     
-        // open the sample copy
-        const scriptCode = fs.readFileSync(sampleFile, "utf-8");    
-        // run the sample file in the browser     
-        await page.evaluate(scriptCode);
+        let fileHash = ""
+        if(!justCheckPage){
+            // open the sample copy
+            const scriptCode = fs.readFileSync(sampleFile, "utf-8");
+            fileHash = sha256(scriptCode)
+    
+            // run the sample file in the browser     
+            await page.evaluate(scriptCode);
+        }
         
         // run the Lure          
         const lure = new Lure(page)
@@ -103,21 +176,28 @@ if (!fs.existsSync(sampleFile)) {
         await lure.start_lure()
 
         logger.debug(`[analysis-debug] Finishing Lure`);
+        logger.info(`[analysis-info] Waiting for ${max_delay} before closing page`);
+        await new Promise(resolve =>
+            setTimeout(resolve, max_delay + 1000)
+        );
         await page.close();
 
-        let iocs_for_analysis: types.IoCsFromAnalysis = {
-            file_hash: sha256(scriptCode),
-            analysis_id: analysisId,
-            iocs: iocs
+        if (!cli_args.dev) {
+            let iocs_for_analysis: types.IoCsFromAnalysis = {
+                file_hash: fileHash,
+                analysis_id: analysisId,
+                iocs: iocs
+            }
+            logger.info(`[analysis-info] found ${iocs_for_analysis.iocs.length} iocs`)
+            await rbmqc.publish(rabbitmq_config.queues.sandbox_iocs_queue.name, iocs_for_analysis)
+            try {
+                fs.rmSync(sampleFile)
+            } catch(e) {
+                logger.error("Could not remove sample file")
+            }
+            await rbmqc.close();
         }
-        logger.info(`[analysis-info] found ${iocs_for_analysis.iocs.length} iocs`)
-        await rbmqc.publish(rabbitmq_config.queues.sandbox_iocs_queue.name, iocs_for_analysis)
-        try {
-            fs.rmSync(sampleFile)
-        } catch(e) {
-            logger.error("Could not remove sample file")
-        }
-        await rbmqc.close();
+
         process.exit(0)
     } catch(err) {
         if (err instanceof SyntaxError) {
@@ -126,9 +206,11 @@ if (!fs.existsSync(sampleFile)) {
             logger.error("Unknown error: ", err);
         }
         await page.close();
-        // should send errors to another queue or anywhere else
-        await rbmqc.publish(rabbitmq_config.queues.sandbox_iocs_queue.name, `error analysing sample: ${err}`)
-        await rbmqc.close();
+        if (!cli_args.dev) {
+            // should send errors to another queue or anywhere else
+            await rbmqc.publish(rabbitmq_config.queues.sandbox_iocs_queue.name, `error analysing sample: ${err}`)
+            await rbmqc.close();
+        }
         process.exit(1)
     }
 })();
